+++ tests/testthat/test-autos.R
@@ -8,7 +8,7 @@ Sys.setenv(testpath = (tmpdir))
 
 system(paste0("cp ", testthat::test_path("man/testdir/*"), " ", tmpdir, " -r"))
 
-envsetup_config <- config::get(
+custom_name <- config::get(
   file = testthat::test_path("man/_envsetup_testthat.yml")
 )
 
@@ -20,7 +20,7 @@ test_that("library returns invisibly",{
 #' @editor Aidan Ceney
 #' @editDate 2022-05-12
 test_that("1.1", {
-  set_autos(envsetup_config$autos)
+  set_autos(custom_name$autos)
   expect_equal(c(test_dev()), c("Test of dev autos"))
 })
 #' @editor Aidan Ceney
@@ -32,9 +32,9 @@ test_that("1.2", {
 #' @editor Aidan Ceney
 #' @editDate 2022-05-12
 test_that("1.3", {
-  envsetup_config_tmp <- envsetup_config
-  envsetup_config_tmp$autos[2] <- envsetup_config$autos[1]
-  set_autos(envsetup_config_tmp$autos)
+  custom_name_tmp <- custom_name
+  custom_name_tmp$autos[2] <- custom_name$autos[1]
+  set_autos(custom_name_tmp$autos)
   expect_equal(
     c(test_dev(), test_prod()),
     c("Test of dev autos", "Test of prod autos")
@@ -46,15 +46,15 @@ test_that("1.3", {
 #' @editor Aidan Ceney
 #' @editDate 2022-05-12
 test_that("1.4", {
-  print(envsetup_config$autos)
-  set_autos(envsetup_config$autos)
+  print(custom_name$autos)
+  set_autos(custom_name$autos)
   expect_equal(mtcars, iris)
 })
 
 #' @editor Aidan Ceney
 #' @editDate 2022-05-12
 test_that("1.5", {
-  set_autos(envsetup_config$autos)
+  set_autos(custom_name$autos)
   expect_equal(test_prod(), "Test of prod autos")
   expect_equal(test_prod2(), "Test of prod autos second")
 })
@@ -62,7 +62,7 @@ test_that("1.5", {
 #' @editor Aidan Ceney
 #' @editDate 2022-05-12
 test_that("1.6", {
-  set_autos(envsetup_config$autos)
+  set_autos(custom_name$autos)
   expect_equal(mtcars, iris)
 })
 
@@ -71,10 +71,10 @@ test_that("1.6", {
 #' @editDate 2022-05-12
 test_that("1.7", {
   Sys.setenv(ENVSETUP_ENVIRON = "QA")
-  set_autos(envsetup_config$autos)
+  set_autos(custom_name$autos)
   expect_error(test_dev())
   Sys.setenv(ENVSETUP_ENVIRON = "PROD")
-  set_autos(envsetup_config$autos)
+  set_autos(custom_name$autos)
   expect_error(test_dev())
 })
 
@@ -82,7 +82,7 @@ test_that("1.7", {
 #' @editDate 2022-05-12
 test_that("1.8", {
   Sys.setenv(ENVSETUP_ENVIRON = "QA")
-  set_autos(envsetup_config$autos)
+  set_autos(custom_name$autos)
   expect_error(test_dev())
   expect_equal(
     c(test_qa(), test_prod()),
@@ -97,4 +97,17 @@ test_that("2.1", {
   detach_autos(c("autos:QA", "autos:PROD"))
   expect_error(test_qa())
   expect_error(test_prod())
+})
+
+
+test_that("the configuration can be named anything and library will
+          reattch the autos correctly", {
+  rprofile(custom_name)
+
+  library("dplyr")
+
+  dplyr_location <- which(search() == "package:dplyr")
+  autos_locatios <- which(grepl("^autos:", search()))
+
+  expect_true(all(dplyr_location > autos_locatios))
 })